Implement shader LEA instruction and improve bindless image load/store (#1355)

+++ Ryujinx.Graphics.Shader/Decoders/BitfieldExtensions.cs
@@ -4,12 +4,12 @@ namespace Ryujinx.Graphics.Shader.Decoders
     {
         public static bool Extract(this int value, int lsb)
         {
-            return ((int)(value >> lsb) & 1) != 0;
+            return ((value >> lsb) & 1) != 0;
         }
 
         public static int Extract(this int value, int lsb, int length)
         {
-            return (int)(value >> lsb) & (int)(uint.MaxValue >> (32 - length));
+            return (value >> lsb) & (int)(uint.MaxValue >> (32 - length));
         }
 
         public static bool Extract(this long value, int lsb)
